requirements.txt
aiogram~=2.18
aioredis<2.0
environs~=9.0
pyowm==3.3.0
py-currency-converter-1.2.0